Components available in BME Design Suite
My first draft. There has to be a better way to save this. Right now it's an arcive of the .LibPkg, PcbLib and SchLib files. I think I read how to compile everything into the PcbLib, but I've failed to do that so far. I have so much effort invested at this point that I don't want to risk loss, so here it all is. Subsequent versions will fix this.